requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
torch
faiss-cpu
langchain
openai
langchain-community
# Dependencies for TangentCFT
gensim==4.3.3
matplotlib==3.10.1
numpy==1.26.4
beautifulsoup4==4.13.3
tqdm==4.67.1
lxml==5.3.1